pull_requests_issues/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
cssselect==1.1.0
cyrusbus==0.1.0
fake-useragent==0.1.11
idna==2.9
lxml==4.5.0
parse==1.15.0
pkg-resources==0.0.0
pyee==7.0.2
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.1
redis==3.5.2
requests==2.23.0
retools==0.4.1
setproctitle==1.1.10
six==1.14.0
soupsieve==2.0
tornado==6.0.4
tqdm==4.46.0
urllib3==1.25.9
w3lib==1.22.0
websockets==8.1
